+++ examples/langchain_rag/.env.example
@@ -2,7 +2,7 @@ DIAL_URL=DIAL_URL
 
 CHAT_MODEL=gpt-4
 EMBEDDINGS_MODEL=text-embedding-ada-002
-API_VERSION=2023-03-15-preview
+API_VERSION=2024-02-01
 
 # Enabling debug logs for OpenAI and Langchain
 OPENAI_LOG=debug
